requirements.txt
flask
textblob
gunicorn
google-generativeai